% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burr.R
\name{burr10}
\alias{burr10}
\alias{dburr10}
\alias{pburr10}
\alias{qburr10}
\alias{rburr10}
\title{Burr type X distribution}
\usage{
dburr10(x, c, k, log = FALSE)

pburr10(q, c, k, lower.tail = TRUE, log.p = FALSE)

qburr10(p, c, k, lower.tail = TRUE, log.p = FALSE)

rburr10(n, c, k)
}
\arguments{
\item{x, q}{Numeric vector of quantiles (>0).}

\item{c}{Positive parameter.}

\item{k}{Positive parameter.}

\item{log}{Logical; if TRUE return log-density.}

\item{lower.tail}{Logical; if TRUE (default) probabilities are \eqn{P[X \le x]}.}

\item{log.p}{Logical; if TRUE return log-probabilities.}

\item{p}{Numeric vector of probabilities between 0 and 1.}

\item{n}{Number of observations.}
}
\description{
Density, distribution function, quantile function and random generation
for the Burr type X distribution on \eqn{x>0}.
}
