% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burr.R
\name{burr3}
\alias{burr3}
\alias{dburr3}
\alias{pburr3}
\alias{qburr3}
\alias{rburr3}
\title{Burr type III distribution}
\usage{
dburr3(x, c, k, s, log = FALSE)

pburr3(q, c, k, s, lower.tail = TRUE, log.p = FALSE)

qburr3(p, c, k, s, lower.tail = TRUE, log.p = FALSE)

rburr3(n, c, k, s)
}
\arguments{
\item{x, q}{Numeric vector of quantiles (>0).}

\item{c, k, s}{Positive parameters.}

\item{log}{Logical; if TRUE return log-density.}

\item{lower.tail}{Logical; if TRUE (default) probabilities are \eqn{P[X \le x]}.}

\item{log.p}{Logical; if TRUE return log-probabilities.}

\item{p}{Numeric vector of probabilities between 0 and 1.}

\item{n}{Number of observations.}
}
\value{
A numeric vector.
}
\description{
Density, distribution function, quantile function and random generation
for the Burr type III distribution on \eqn{x>0}.
}
