% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burr.R
\name{meanburr12}
\alias{meanburr12}
\title{Mean of the Burr type XII distribution}
\usage{
meanburr12(c, k, s)
}
\arguments{
\item{c, k, s}{Positive parameters.}
}
\value{
A numeric value (may be \code{Inf} if the mean does not exist).
}
\description{
Mean for the Burr type XII distribution (this parameterization requires \code{k > 1}).
The mean exists only if \code{k > 1 + 1/c}.
}
