% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burr.R
\name{varburr12}
\alias{varburr12}
\title{Variance of the Burr type XII distribution}
\usage{
varburr12(c, k, s)
}
\arguments{
\item{c, k, s}{Positive parameters.}
}
\value{
A numeric value (may be \code{Inf} if the variance does not exist).
}
\description{
Variance for the Burr type XII distribution (this parameterization requires \code{k > 1}).
The variance exists only if \code{k > 1 + 2/c}.
}
